Attempt at creating the "playlist" functionality
The idea is that, upon scanning the folder, the properties of the script should be updated, adding more properties to it, one for each song in the scanned folder, that allows to set an order of reproduction for each song. However, I can't for the life of me add these new properties. Here I try to re-enter the script_update function since it's the only place where properties can be added seemingly, but apparently that only works on program startup. So, yeah. Completely blocked here.